batch: Extract ownership translation
Batch ownership contains claim metadata, normalization, merge, storage
acquisition, and selected-line translation helpers.

Selected-line translation scans LineEntry objects, builds absence content,
and detects stale batch sources, which is a separate flow from the
ownership data model.

This commit moves selected-line translation into
batch.ownership_translation and updates production callers to import that
owner.

Subsequent commits will move tests and architecture docs to the translation
boundary.
