+++ src/git_stage_batch/batch/source_refresh.py
@@ -25,8 +25,10 @@ from .lineage import BatchSourceLineage
 from .match import match_lines
 from .ownership import (
     BatchOwnership,
-    detect_stale_batch_source_for_selection,
     merge_batch_ownership,
+)
+from .ownership_translation import (
+    detect_stale_batch_source_for_selection,
     translate_hunk_selection_to_batch_ownership,
     translate_lines_to_batch_ownership,
 )
