+++ src/git_stage_batch/data/consumed_selections.py
@@ -7,8 +7,10 @@ from typing import Any
 
 from ..batch.ownership import (
     BatchOwnership,
-    detect_stale_batch_source_for_selection,
     merge_batch_ownership,
+)
+from ..batch.ownership_translation import (
+    detect_stale_batch_source_for_selection,
     translate_lines_to_batch_ownership,
 )
 from ..batch.source_advancement import advance_batch_source_for_file_with_provenance
